backendapp/requirements.txt
fastapi>=0.110.0
uvicorn[standard]
pydantic
litellm
langfuse
ragas>=0.4.0
rapidfuzz
sacrebleu
rouge-score
python-dotenv
pytest
httpx
